crypto: move make_nonce to crypto/mod.rs
